extract normalizedDiagnostics
